+++ src/lib.rs
@@ -70,6 +70,7 @@ detected.
   * [Microsoft Visio Drawing (VSD)](`FileFormat::MicrosoftVisioDrawing`)
   * [Microsoft Word Document (DOC)](`FileFormat::MicrosoftWordDocument`)
   * [SolidWorks Assembly (SLDASM)](`FileFormat::SolidworksAssembly`)
+  * [SolidWorks Drawing (SLDDRW)](`FileFormat::SolidworksDrawing`)
   * [SolidWorks Part (SLDPRT)](`FileFormat::SolidworksPart`)
   * [StarCalc (SDC)](`FileFormat::Starcalc`)
   * [StarChart (SDS)](`FileFormat::Starchart`)