+++ tests/model.rs
@@ -175,6 +175,13 @@ fn test_solidworks_assembly() {
 
 #[cfg(feature = "reader-cfb")]
 #[test]
+fn test_solidworks_drawing() {
+    let fmt = FileFormat::from_file("fixtures/model/sample.slddrw").unwrap();
+    assert_eq!(fmt, FileFormat::SolidworksDrawing);
+}
+
+#[cfg(feature = "reader-cfb")]
+#[test]
 fn test_solidworks_part() {
     let fmt = FileFormat::from_file("fixtures/model/sample.sldprt").unwrap();
     assert_eq!(fmt, FileFormat::SolidworksPart);